requirements.txt
Flask
Flask-SQLAlchemy
Flask-RQ
Flask-Migrate
https://github.com/hasgeek/coaster/archive/master.zip
https://github.com/hasgeek/baseframe/archive/master.zip
langid
rq
